% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{.compute_allowed_genotypes}
\alias{.compute_allowed_genotypes}
\title{Compute allowed offspring genotypes per locus}
\usage{
.compute_allowed_genotypes(sire_geno, dam_geno)
}
\arguments{
\item{sire_geno}{Integer vector of sire dosages (0/1/2/NA).}

\item{dam_geno}{Integer vector of dam dosages (0/1/2/NA).}
}
\value{
A logical matrix (n_loci x 3) where columns correspond to
dosage 0, 1, 2 indicating whether each is an allowed offspring genotype.
}
\description{
Compute allowed offspring genotypes per locus
}
\keyword{internal}
