% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{.is_compatible}
\alias{.is_compatible}
\title{Check compatibility of individual genotypes with allowed set}
\usage{
.is_compatible(ind_geno, allowed)
}
\arguments{
\item{ind_geno}{Integer vector of individual dosages.}

\item{allowed}{Logical matrix (n_loci x 3).}
}
\value{
Logical vector; TRUE if compatible.
}
\description{
Check compatibility of individual genotypes with allowed set
}
\keyword{internal}
