% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.R
\name{print.aapa_result}
\alias{print.aapa_result}
\title{Print method for aapa_result}
\usage{
\method{print}{aapa_result}(x, ...)
}
\arguments{
\item{x}{An \code{aapa_result} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for aapa_result
}
\seealso{
Other assignment: 
\code{\link{aapa_assign}()},
\code{\link{summary.aapa_result}()}
}
\concept{assignment}
