% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.R
\name{summary.aapa_result}
\alias{summary.aapa_result}
\title{Summary method for aapa_result}
\usage{
\method{summary}{aapa_result}(object, ...)
}
\arguments{
\item{object}{An \code{aapa_result} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Summary method for aapa_result
}
\seealso{
Other assignment: 
\code{\link{aapa_assign}()},
\code{\link{print.aapa_result}()}
}
\concept{assignment}
